Reorganize account subpackage
Fix account GET bug
