+++ tests/account/test_db.py
@@ -1,6 +1,7 @@
 import pytest
 from aiohttp.test_utils import make_mocked_coro
 import virtool.account.db
+import virtool.account.utils
 
 
 def test_compose_password_update(mocker, static_time):
@@ -40,25 +41,6 @@ async def test_get_alternate_id(existing, expected, dbi):
     assert await virtool.account.db.get_alternate_id(dbi, "foo") == expected
 
 
-def test_generate_api_key(mocker):
-    """
-    Test that API keys are generated using UUID4 and that :func:`generate_api_key()` returns the raw and hashed version
-    of the key. Hashing is done through a call to :func:`hash_api_key`.
-
-    """
-    class MockUUID4:
-        hex = "foo"
-
-    m_uuid = mocker.patch("uuid.uuid4", return_value=MockUUID4())
-    m_hash_api_key = mocker.patch("virtool.users.utils.hash_api_key", return_value="bar")
-
-    assert virtool.account.db.generate_api_key() == ("foo", "bar")
-
-    assert m_uuid.called
-
-    m_hash_api_key.assert_called_with("foo")
-
-
 @pytest.mark.parametrize("administrator", [True, False], ids=["administrator", "limited"])
 @pytest.mark.parametrize("has_permission", [True, False], ids=["has permission", "missing permission"])
 async def test_create_api_key(administrator, has_permission, mocker, dbi, static_time):
@@ -67,7 +49,7 @@ async def test_create_api_key(administrator, has_permission, mocker, dbi, static
 
     """
     m_get_alternate_id = mocker.patch("virtool.account.db.get_alternate_id", make_mocked_coro("foo_0"))
-    m_generate_api_key = mocker.patch("virtool.account.db.generate_api_key", return_value=("bar", "baz"))
+    m_generate_api_key = mocker.patch("virtool.account.utils.generate_api_key", return_value=("bar", "baz"))
 
     groups = [
         "technicians",
